backend/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
flask-cors==5.0.0
gunicorn==23.0.0
